[CSL-885] Separate Network.Retrieval and Network.Logic
