+++ src/Pos/Block/Network/Listeners.hs
@@ -7,32 +7,32 @@ module Pos.Block.Network.Listeners
        , blockStubListeners
        ) where
 
-import           Data.Reflection             (reify)
-import           Data.Tagged                 (Tagged, proxy, unproxy)
-import           Formatting                  (build, sformat, (%))
-import qualified Node                        as N
-import           Serokell.Data.Memory.Units  (Byte)
-import           Serokell.Util.Text          (listJson)
-import           System.Wlog                 (WithLogger, logDebug, logWarning,
-                                              modifyLoggerName)
+import           Data.Reflection            (reify)
+import           Data.Tagged                (Tagged, proxy, unproxy)
+import           Formatting                 (build, sformat, (%))
+import qualified Node                       as N
+import           Serokell.Data.Memory.Units (Byte)
+import           Serokell.Util.Text         (listJson)
+import           System.Wlog                (WithLogger, logDebug, logWarning,
+                                             modifyLoggerName)
 import           Universum
 
-import           Pos.Binary.Communication    ()
-import           Pos.Block.Logic             (getHeadersFromToIncl)
-import           Pos.Block.Network.Announce  (handleHeadersCommunication)
-import           Pos.Block.Network.Retrieval (handleUnsolicitedHeaders)
-import           Pos.Block.Network.Types     (MsgBlock (..), MsgGetBlocks (..),
-                                              MsgGetHeaders (..), MsgHeaders (..))
-import           Pos.Communication.Limits    (recvLimited, reifyMsgLimit)
-import           Pos.Communication.Protocol  (ConversationActions (..), HandlerSpec (..),
-                                              ListenerSpec (..), OutSpecs, listenerConv,
-                                              mergeLs, messageName)
-import           Pos.Communication.Util      (stubListenerConv)
-import qualified Pos.DB.Block                as DB
-import           Pos.DB.Error                (DBError (DBMalformed))
-import           Pos.Ssc.Class               (SscHelpersClass, SscWorkersClass)
-import           Pos.Util                    (NewestFirst (..))
-import           Pos.WorkMode                (WorkMode)
+import           Pos.Binary.Communication   ()
+import           Pos.Block.Logic            (getHeadersFromToIncl)
+import           Pos.Block.Network.Announce (handleHeadersCommunication)
+import           Pos.Block.Network.Logic    (handleUnsolicitedHeaders)
+import           Pos.Block.Network.Types    (MsgBlock (..), MsgGetBlocks (..),
+                                             MsgGetHeaders (..), MsgHeaders (..))
+import           Pos.Communication.Limits   (recvLimited, reifyMsgLimit)
+import           Pos.Communication.Protocol (ConversationActions (..), HandlerSpec (..),
+                                             ListenerSpec (..), OutSpecs, listenerConv,
+                                             mergeLs, messageName)
+import           Pos.Communication.Util     (stubListenerConv)
+import qualified Pos.DB.Block               as DB
+import           Pos.DB.Error               (DBError (DBMalformed))
+import           Pos.Ssc.Class              (SscHelpersClass, SscWorkersClass)
+import           Pos.Util                   (NewestFirst (..))
+import           Pos.WorkMode               (WorkMode)
 
 blockListeners
     :: (SscWorkersClass ssc, WorkMode ssc m)
